[ctx_prof] Move the "from json" logic more centrally to reuse it from test.
